Integracion lat-a-utf-8 en comun
Se comenzo la integracion del codigo para la traduccion de latin1 a
utf-8 en comun, de forma que facilmente pueda ser llamado desde otros archivos
